Statusbar related code moved to separate namespace for clarity
