+++ ReasonCodeExample.XPathTools.Tests/ReasonCodeExample.XPathTools.Tests.csproj
@@ -163,8 +163,8 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="VisualStudioIntegration\StatusbarAdapterTests.cs" />
-    <Compile Include="VisualStudioIntegration\XPathToolsPackageTests.cs" />
+    <Compile Include="Statusbar\StatusbarAdapterTests.cs" />
+    <Compile Include="RegistryTests.cs" />
     <Compile Include="Writers\XPathWriterFactoryTests.cs" />
     <Compile Include="XmlStringExtensions.cs" />
     <Compile Include="VisualStudioIntegration\CaretPositionLineInfoTests.cs" />
