+++ ReasonCodeExample.XPathTools.Tests/RegistryTests.cs
@@ -3,11 +3,12 @@ using System.ComponentModel.Design;
 using Microsoft.VisualStudio.Shell.Interop;
 using NSubstitute;
 using NUnit.Framework;
+using ReasonCodeExample.XPathTools.Statusbar;
 using ReasonCodeExample.XPathTools.VisualStudioIntegration;
 
-namespace ReasonCodeExample.XPathTools.Tests.VisualStudioIntegration
+namespace ReasonCodeExample.XPathTools.Tests
 {
-    public class XPathToolsPackageTests
+    public class RegistryTests
     {
         [TestCase(XPathFormat.Generic, "<a><b id='hello'><c/></b></a>", 19, "/a/b[@id='hello']/c")]
         [TestCase(XPathFormat.Absolute, "<a id='world'><b><c/></b></a>", 19, "/a[1][@id='world']/b[1]/c[1]")]
@@ -15,8 +16,6 @@ namespace ReasonCodeExample.XPathTools.Tests.VisualStudioIntegration
         {
             // Arrange
             var serviceContainer = new ServiceContainer();
-
-            var package = new XPathToolsPackage(serviceContainer);
 
             // Set up the configuration to a default XPathFormat
             var configuration = Substitute.For<IConfiguration>();
@@ -28,7 +27,7 @@ namespace ReasonCodeExample.XPathTools.Tests.VisualStudioIntegration
             // Initialize all services, incl. the statusbar adapter
             var menuCommandService = Substitute.For<IMenuCommandService>();
             var statusbar = Substitute.For<IVsStatusbar>();
-            package.Initialize(configuration, menuCommandService, statusbar);
+            //registry.Initialize(configuration, menuCommandService, statusbar);
 
             // Set up the XML repo initialized by the package
             var xmlRepository = serviceContainer.Get<XmlRepository>();
